refactor(config): Remove the config file
- Added a main_test.go to store the [TestMain function](https://golang.org/pkg/testing/#hdr-Main) for this package for running
`config.InitConfig()`.
- Moved some mock functions frpm config_test.go into main_test.go
