+++ pkg/app/app.go
@@ -55,7 +55,6 @@ func (a *AppBuilder) WithRetrieval() *AppBuilder {
 
 	a.components = append(a.components, newExpirationWorker(a.database))
 
-	a.servlets = append(a.servlets, server.NewConfigServlet())
 	a.servlets = append(a.servlets, server.NewRetrieveServlet(a.database, retrieval.NewAuthenticator(), retrieval.NewSigner()))
 
 	return a
